Federated Machine Learning XGboost/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "xgboost_quickstart"
version = "1.0.0"
description = "Federated Learning with XGBoost and Flower (Quickstart Example)"
license = "Apache-2.0"
dependencies = [
    "flwr-nightly[simulation]==1.11.0.dev20240826",
    "flwr-datasets>=0.3.0",
    "xgboost>=2.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.flwr.app]
publisher = "flwrlabs"

[tool.flwr.app.components]
serverapp = "my_server:app"
clientapp = "my_client:app"

[tool.flwr.app.config]
# ServerApp
num-server-rounds = 10
fraction-fit = 0.1
fraction-evaluate = 0.1

# ClientApp
local-epochs = 1
params.objective = "binary:logistic"
params.eta = 0.1 # Learning rate
params.max-depth = 8
params.eval-metric = "auc"
params.nthread = 16
params.num-parallel-tree = 1
params.subsample = 1
params.tree-method = "hist"

[tool.flwr.federations]
default = "local-simulation"

[tool.flwr.federations.local-simulation]
options.num-supernodes = 2